Fix bogus information in dircache
The code was only zeroing the first 32 *bytes* of the newly-allocated space
after a successful `realloc()`, rather than all 32 *entries' worth* of space.
Switched to `calloc()` plus `memcpy()` to make the intent more clear and
ensure zeroing of the full allocation, since zeroing allocations are
effectively free on modern systems.